Fairness Solver Started
Wrote an IP to solve for the highest (weighted) percentage of vertices
in a set of "special vertices" (e.g., highly-sensitized vertices) over
the set of all possible matches in a pool.

Revamped the Solvers so now CPLEXSolver serves as a superclass for all
CPLEX-based solvers, does initialization and basic solution parsing.
CPLEXSolver still a subclass of Solver.

Something is wrong with the chain generation part of CycleGenerator.